Move Drawer into backend
The current `Drawer` class is very X11-specific in its implementation,
but the API it exposes to widgets etc needs to be consistent in other
implementations. To this end, this commit splits the current
implementation in two parts:

 - A generic base primarily concerned with defining methods and cairo
   logic in the base backend.
 - And the X11-specific layer on top of this that lets the drawer paint
   to X11 windows, in x11/drawer.py.

In order to create drawers -- which are bound to Internal windows -- is
via a new method on Internal windows: `Internal.create_drawer(width,
height)`. This no longer requires modules to import drawer.py to
instantiate Drawers.

The cairocffi parts remain in the base class because the Wayland
implementation will also use cairocffi, and this avoids duplicate code
or abstracting for abstracting's sake.
