+++ test/widgets/test_mpris2widget.py
@@ -59,12 +59,6 @@ class obj:  # noqa: N801
         self.value = value
 
 
-class FakeWindow:
-    class _NestedWindow:
-        wid = 10
-    window = _NestedWindow()
-
-
 # Creates a mock message body containing both metadata and playback status
 def metadata_and_status(status):
     return MockMessage(body=(
@@ -119,10 +113,10 @@ def patched_module(monkeypatch):
     return mpris2widget
 
 
-def test_mpris2_signal_handling(fake_qtile, patched_module):
+def test_mpris2_signal_handling(fake_qtile, patched_module, fake_window):
     mp = patched_module.Mpris2(scroll_chars=20, scroll_wait_intervals=5)
     fakebar = Bar([mp], 24)
-    fakebar.window = FakeWindow()
+    fakebar.window = fake_window
     fakebar.width = 10
     fakebar.height = 10
     fakebar.draw = no_op
@@ -182,10 +176,10 @@ def test_mpris2_signal_handling(fake_qtile, patched_module):
     assert info["isplaying"]
 
 
-def test_mpris2_custom_stop_text(fake_qtile, patched_module):
+def test_mpris2_custom_stop_text(fake_qtile, patched_module, fake_window):
     mp = patched_module.Mpris2(stop_pause_text="Test Paused")
     fakebar = Bar([mp], 24)
-    fakebar.window = FakeWindow()
+    fakebar.window = fake_window
     fakebar.width = 10
     fakebar.height = 10
     fakebar.draw = no_op
@@ -203,10 +197,10 @@ def test_mpris2_custom_stop_text(fake_qtile, patched_module):
     assert mp.displaytext == "Test Paused"
 
 
-def test_mpris2_no_metadata(fake_qtile, patched_module):
+def test_mpris2_no_metadata(fake_qtile, patched_module, fake_window):
     mp = patched_module.Mpris2(stop_pause_text="Test Paused")
     fakebar = Bar([mp], 24)
-    fakebar.window = FakeWindow()
+    fakebar.window = fake_window
     fakebar.width = 10
     fakebar.height = 10
     fakebar.draw = no_op
@@ -218,12 +212,12 @@ def test_mpris2_no_metadata(fake_qtile, patched_module):
     assert mp.displaytext == "No metadata for current track"
 
 
-def test_mpris2_no_scroll(fake_qtile, patched_module):
+def test_mpris2_no_scroll(fake_qtile, patched_module, fake_window):
     # If no scrolling, then the update function creates the text to display
     # and draws the bar.
     mp = patched_module.Mpris2(scroll_chars=None)
     fakebar = Bar([mp], 24)
-    fakebar.window = FakeWindow()
+    fakebar.window = fake_window
     fakebar.width = 10
     fakebar.height = 10
     fakebar.draw = no_op
@@ -238,10 +232,10 @@ def test_mpris2_no_scroll(fake_qtile, patched_module):
     assert mp.text == "Paused: Never Gonna Give You Up - Whenever You Need Somebody - Rick Astley"
 
 
-def test_mpris2_clear_after_scroll(fake_qtile, patched_module):
+def test_mpris2_clear_after_scroll(fake_qtile, patched_module, fake_window):
     mp = patched_module.Mpris2(scroll_chars=60, scroll_wait_intervals=2)
     fakebar = Bar([mp], 24)
-    fakebar.window = FakeWindow()
+    fakebar.window = fake_window
     fakebar.width = 10
     fakebar.height = 10
     fakebar.draw = no_op
